+++ psengine/playbook_alerts/markdown/__init__.py
@@ -10,4 +10,3 @@
 # for having all necessary licenses, permissions, rights, and/or consents to any data        #
 # accessed from any third party API.                                                         #
 ##############################################################################################
-